Add tests for PIs, cast the index values to ints in diagrams/_utils

+++ gtda/diagrams/_utils.py
@@ -40,7 +40,7 @@ def _sort(Xs):
 
 def _sample_image(image, sampled_diag):
     unique, counts = np.unique(sampled_diag, axis=0, return_counts=True)
-    unique = tuple(tuple(row) for row in unique.T)
+    unique = tuple(tuple(row) for row in unique.astype(np.int).T)
     image[unique] = counts
